# docs/swagger.yaml
basePath: /
definitions:
  controller.LoginPayload:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  controller.UserPasswordPayload:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
      userID:
        type: integer
    type: object
  controller.UserRolePayload:
    properties:
      newRole:
        type: string
      userID:
        type: integer
    type: object
...
  model.Response:
    properties:
      code:
        type: string
      data:
        type: object
      message:
        type: string
      success:
        type: boolean
    type: object
...
  model.UserPayload:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Swagger
  version: "1.0"
paths:
...
  /user/{uid}:
    get:
      consumes:
      - application/json
      description: gets user info
      parameters:
      - description: User ID
        in: path
        name: uid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: gets user info
      tags:
      - user-manager-apis
  /user/all:
    get:
      consumes:
      - application/json
      description: get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: get all users
      tags:
      - user-manager-apis
  /user/create:
    post:
      consumes:
      - application/json
      description: creates new user
      parameters:
      - description: User information
        in: body
        name: UserInfo
        required: true
        schema:
          $ref: '#/definitions/model.UserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: creates new user
      tags:
      - user-manager-apis
  /user/delete/{uid}:
    delete:
      consumes:
      - application/json
      description: delete user
      parameters:
      - description: User ID
        in: path
        name: uid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: delete user
      tags:
      - user-manager-apis
  /user/login:
    post:
      consumes:
      - application/json
      description: log user in
      parameters:
      - description: username & password
        in: body
        name: LoginPayload
        required: true
        schema:
          $ref: '#/definitions/controller.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: log user in
      tags:
      - user-manager-apis
...
  /user/modify/pwd:
    put:
      consumes:
      - application/json
      description: change password
      parameters:
      - description: User and password info
        in: body
        name: UserPasswordInfo
        required: true
        schema:
          $ref: '#/definitions/controller.UserPasswordPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: change password
      tags:
      - user-manager-apis
  /user/modify/role:
    put:
      consumes:
      - application/json
      description: change role
      parameters:
      - description: UserID and role
        in: body
        name: UserRoleInfo
        required: true
        schema:
          $ref: '#/definitions/controller.UserRolePayload'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: change role
      tags:
      - user-manager-apis
  /user/reset/{uid}:
    put:
      consumes:
      - application/json
      description: reset password
      parameters:
      - description: User ID
        in: path
        name: uid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: reset password
      tags:
      - user-manager-apis
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"